s3_client: isolate `retryable_http_client`
Relocated `retryable_http_client` into its own dedicated file for improved clarity and maintainability.
