+++ configure.py
@@ -1031,6 +1031,7 @@ scylla_core = (['message/messaging_service.cc',
                 'utils/gz/crc_combine_table.cc',
                 'utils/s3/aws_error.cc',
                 'utils/s3/client.cc',
+                'utils/s3/retryable_http_client.cc',
                 'utils/s3/retry_strategy.cc',
                 'utils/s3/credentials_providers/aws_credentials_provider.cc',
                 'utils/s3/credentials_providers/environment_aws_credentials_provider.cc',
